refactor(checklist): move formatChecklist to checklist-format
