+++ src/checklist-contract.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "bun:test";
-import { checklistMarker, checklistProgress, formatChecklist } from "./checklist-contract";
+import { checklistMarker, checklistProgress } from "./checklist-contract";
 
 describe("checklistMarker", () => {
   test("returns correct markers", () => {
@@ -41,31 +41,5 @@ describe("checklistProgress", () => {
 
   test("handles empty list", () => {
     expect(checklistProgress([])).toEqual({ done: 0, total: 0 });
-  });
-});
-
-describe("formatChecklist", () => {
-  test("returns header with progress and sorted lines", () => {
-    const result = formatChecklist({
-      groupId: "g1",
-      groupTitle: "Build",
-      items: [
-        { id: "s2", label: "test", status: "in_progress", order: 1 },
-        { id: "s1", label: "lint", status: "done", order: 0 },
-        { id: "s3", label: "deploy", status: "pending", order: 2 },
-      ],
-    });
-    expect(result.header).toBe("Build (1/3)");
-    expect(result.lines).toEqual(["● lint", "◐ test", "○ deploy"]);
-  });
-
-  test("handles single item", () => {
-    const result = formatChecklist({
-      groupId: "g1",
-      groupTitle: "Quick",
-      items: [{ id: "s1", label: "do it", status: "pending", order: 0 }],
-    });
-    expect(result.header).toBe("Quick (0/1)");
-    expect(result.lines).toEqual(["○ do it"]);
   });
 });
