+++ src/checklist-contract.ts
@@ -37,12 +37,3 @@ export function checklistProgress(items: ChecklistItem[]): { done: number; total
     total: items.length,
   };
 }
-
-export function formatChecklist(output: ChecklistOutput): { header: string; lines: string[] } {
-  const sorted = [...output.items].sort((a, b) => a.order - b.order);
-  const { done, total } = checklistProgress(sorted);
-  return {
-    header: `${output.groupTitle} (${done}/${total})`,
-    lines: sorted.map((item) => `${checklistMarker(item.status)} ${item.label}`),
-  };
-}
